template form modal is ok
